+++ packages/react-components/react-combobox/src/components/Combobox/renderCombobox.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Portal } from '@fluentui/react-portal';
 import { getSlots } from '@fluentui/react-utilities';
-import type { ComboboxContextValues, ComboboxState, ComboboxSlots } from './Combobox.types';
+import type { ComboboxState, ComboboxSlots, ComboboxContextValues } from './Combobox.types';
 import { ComboboxContext } from '../../contexts/ComboboxContext';
 
 /**
@@ -11,7 +11,7 @@ export const renderCombobox_unstable = (state: ComboboxState, contextValues: Com
   const { slots, slotProps } = getSlots<ComboboxSlots>(state);
 
   const listbox = <slots.listbox {...slotProps.listbox}>{slotProps.root.children}</slots.listbox>;
-  const popup = state.inline ? listbox : <Portal>{listbox}</Portal>;
+  const popup = state.inlinePopup ? listbox : <Portal>{listbox}</Portal>;
 
   return (
     <slots.root {...slotProps.root}>