+++ packages/react-components/theme-designer/src/components/Nav/Nav.tsx
@@ -1,12 +1,14 @@
 import * as React from 'react';
 import { makeStyles, mergeClasses } from '@griffel/react';
-import { Text, BrandVariants, FluentProvider, webDarkTheme } from '@fluentui/react-components';
+import { Text, BrandVariants, FluentProvider, webDarkTheme, Theme } from '@fluentui/react-components';
 import { CircleRegular, ChevronRightRegular } from '@fluentui/react-icons';
 import { ExportButton } from '../ExportButton/ExportButton';
 
 export interface NavProps {
   className?: string;
   brand: BrandVariants;
+  isDark: boolean;
+  overrides: Partial<Theme>;
 }
 
 const useStyles = makeStyles({
@@ -55,7 +57,7 @@ export const Nav: React.FC<NavProps> = props => {
         UI Colors <ChevronRightRegular /> New palette
       </div>
       <Name />
-      <ExportButton brand={props.brand} />
+      <ExportButton brand={props.brand} isDark={props.isDark} overrides={props.overrides} />
     </FluentProvider>
   );
 };