+++ packages/azure-themes/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/azure-themes
 
-This log was last generated on Mon, 20 Jun 2022 07:45:13 GMT and should not be manually modified.
+This log was last generated on Fri, 24 Jun 2022 07:43:32 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.2.51](https://github.com/microsoft/fluentui/tree/@fluentui/azure-themes_v8.2.51)
+
+Fri, 24 Jun 2022 07:43:32 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/azure-themes_v8.2.50..@fluentui/azure-themes_v8.2.51)
+
+### Patches
+
+- Bump @fluentui/react to v8.77.1 ([PR #23440](https://github.com/microsoft/fluentui/pull/23440) by beachball)
 
 ## [8.2.50](https://github.com/microsoft/fluentui/tree/@fluentui/azure-themes_v8.2.50)